feat: Move person files to new package
All the data related to person-part of the portal should go to the same
package (similar to the new analysis package), this is the first step to
separate all between ocds (action) and person.
