+++ portal/src/person/pages/PersonDetailPage.css
@@ -9,7 +9,7 @@
 .data-box {
   margin: 10px 0;
   border-radius: 10px;
-  box-shadow: 0px 0px 10px rgba(227, 232, 236, 1);
+  box-shadow: 0 0 10px rgba(227, 232, 236, 1);
 }
 
 .data-box .ant-card-head {
