+++ portal/src/person/pages/PersonDetailPage.tsx
@@ -1,32 +1,32 @@
 import * as React from 'react';
 import {useEffect, useMemo, useState} from 'react';
 import {Avatar, Card, Col, Layout, Row, Space, Tooltip, Typography} from 'antd';
-import {Header} from '../components/layout/Header';
+import {Header} from '../../components/layout/Header';
 import './PersonDetailPage.css'
-import Footer from '../components/layout/Footer';
-import {ReactComponent as Pytyvo} from '../assets/logos/pytyvo.svg';
-import {ReactComponent as Nangareko} from '../assets/logos/nangareko.svg';
-import {ReactComponent as Ande} from '../assets/logos/ande.svg';
-import {ReactComponent as HaciendaIcon} from '../assets/logos/hacienda.svg';
-import {ReactComponent as Sfp} from '../assets/logos/sfp.svg';
-import {ReactComponent as Ddjj} from '../assets/logos/ddjj.svg';
-import {ReactComponent as Aqe} from '../assets/logos/a_quienes_elegimos.svg';
-import {ReactComponent as PoliciaNacional} from '../assets/logos/policia_nacional.svg';
+import Footer from '../../components/layout/Footer';
+import {ReactComponent as Pytyvo} from '../../assets/logos/pytyvo.svg';
+import {ReactComponent as Nangareko} from '../../assets/logos/nangareko.svg';
+import {ReactComponent as Ande} from '../../assets/logos/ande.svg';
+import {ReactComponent as HaciendaIcon} from '../../assets/logos/hacienda.svg';
+import {ReactComponent as Sfp} from '../../assets/logos/sfp.svg';
+import {ReactComponent as Ddjj} from '../../assets/logos/ddjj.svg';
+import {ReactComponent as Aqe} from '../../assets/logos/a_quienes_elegimos.svg';
+import {ReactComponent as PoliciaNacional} from '../../assets/logos/policia_nacional.svg';
 import Icon from '@ant-design/icons';
-import {SimpleApi} from '../SimpleApi';
+import {SimpleApi} from '../../SimpleApi';
 import {useParams} from 'react-router-dom';
-import {Affidavit, AnalysisSearchResult, Authorities, Hacienda, LocalSearchResult} from '../Model';
-import {formatMoney, getInitials} from '../formatters';
-import {AQECard} from '../components/person_cards/AQE';
-import {TSJECard} from '../components/person_cards/TSJE';
-import {DDJJCard} from "../components/person_cards/DDJJ";
-import {Charge, ChargeCard} from "../components/person_cards/Charge";
-import {HaciendaCard} from "../components/person_cards/Hacienda";
-import {SFPCard} from '../components/person_cards/SFP';
-import {COLOR_GREY, COLOR_ORANGE} from "../Constants";
+import {Affidavit, AnalysisSearchResult, Authorities, Hacienda, LocalSearchResult} from '../../Model';
+import {formatMoney, getInitials} from '../../formatters';
+import {AQECard} from '../widgets/AQE';
+import {TSJECard} from '../widgets/TSJE';
+import {DDJJCard} from "../widgets/DDJJ";
+import {Charge, ChargeCard} from "../widgets/Charge";
+import {HaciendaCard} from "../widgets/Hacienda";
+import {SFPCard} from '../widgets/SFP';
+import {COLOR_GREY, COLOR_ORANGE} from "../../Constants";
 import {useMediaQuery} from '@react-hook/media-query';
-import {SOURCE_NAME_MAP} from "./PersonSearchPage";
-import {fixName} from '../nameUtils';
+import {SOURCE_NAME_MAP} from "../../pages/PersonSearchPage";
+import {fixName} from '../../nameUtils';
 import useMetaTags from 'react-metatags-hook';
 
 export function PersonDetailPage() {
